+++ cmd/ccd/ast/ast.go
@@ -20,6 +20,12 @@ type AlterProjectStmt struct {
 
 func (*AlterProjectStmt) node() {}
 
+type ArchiveProjectStmt struct {
+	Project string
+}
+
+func (*ArchiveProjectStmt) node() {}
+
 type CreateFundStmt struct {
 	Fund string
 }
@@ -90,9 +96,10 @@ type SelectStmt struct {
 func (*SelectStmt) node() {}
 
 type ShowStmt struct {
-	Type string
-	Name string
-	In   string
+	Type     string
+	Name     string
+	In       string
+	Archived bool
 }
 
 func (*ShowStmt) node() {}