+++ cmd/ccd/parser/gram.go
@@ -23,46 +23,49 @@ const ADD = 57349
 const ALL = 57350
 const ALTER = 57351
 const AND = 57352
-const ASC = 57353
-const BY = 57354
-const CREATE = 57355
-const COUNT = 57356
-const DELETE = 57357
-const DESC = 57358
-const DROP = 57359
-const ENCRYPTED = 57360
-const FILTER = 57361
-const FROM = 57362
-const FUND = 57363
-const ILIKE = 57364
-const IN = 57365
-const INFO = 57366
-const INSERT = 57367
-const INTO = 57368
-const LIKE = 57369
-const LIMIT = 57370
-const NOT = 57371
-const NULL = 57372
-const OFFSET = 57373
-const OR = 57374
-const ORDER = 57375
-const PASSWORD = 57376
-const PROJECT = 57377
-const PROPERTY = 57378
-const PING = 57379
-const RETRIEVE = 57380
-const SELECT = 57381
-const SET = 57382
-const SHOW = 57383
-const TAG = 57384
-const TO = 57385
-const UPDATE = 57386
-const USER = 57387
-const WHERE = 57388
-const WITH = 57389
-const IDENT = 57390
-const NUMBER = 57391
-const SLITERAL = 57392
+const ARCHIVE = 57353
+const ARCHIVED = 57354
+const ASC = 57355
+const BY = 57356
+const CREATE = 57357
+const COUNT = 57358
+const DELETE = 57359
+const DESC = 57360
+const DROP = 57361
+const ENCRYPTED = 57362
+const FILTER = 57363
+const FROM = 57364
+const FUND = 57365
+const ILIKE = 57366
+const IN = 57367
+const INFO = 57368
+const INSERT = 57369
+const INTO = 57370
+const LIKE = 57371
+const LIMIT = 57372
+const NOT = 57373
+const NULL = 57374
+const OFFSET = 57375
+const OR = 57376
+const ORDER = 57377
+const PASSWORD = 57378
+const PROJECT = 57379
+const PROJECTS = 57380
+const PROPERTY = 57381
+const PING = 57382
+const RETRIEVE = 57383
+const SELECT = 57384
+const SET = 57385
+const SHOW = 57386
+const TAG = 57387
+const TO = 57388
+const UPDATE = 57389
+const USER = 57390
+const WHERE = 57391
+const WITH = 57392
+const IDENT = 57393
+const NUMBER = 57394
+const SLITERAL = 57395
 
 var yyToknames = [...]string{
 	"$end",
@@ -75,6 +78,8 @@ var yyToknames = [...]string{
 	"ALL",
 	"ALTER",
 	"AND",
+	"ARCHIVE",
+	"ARCHIVED",
 	"ASC",
 	"BY",
 	"CREATE",
@@ -100,6 +105,7 @@ var yyToknames = [...]string{
 	"ORDER",
 	"PASSWORD",
 	"PROJECT",
+	"PROJECTS",
 	"PROPERTY",
 	"PING",
 	"RETRIEVE",
@@ -139,130 +145,134 @@ var yyExca = [...]int8{
 
 const yyPrivate = 57344
 
-const yyLast = 208
+const yyLast = 217
 
 var yyAct = [...]uint8{
-	100, 148, 95, 91, 89, 147, 96, 92, 62, 74,
-	122, 121, 167, 151, 165, 166, 21, 166, 106, 124,
-	22, 123, 23, 64, 24, 186, 185, 184, 44, 48,
-	50, 25, 26, 52, 53, 54, 55, 56, 57, 58,
-	59, 109, 98, 61, 30, 98, 27, 46, 28, 108,
-	67, 29, 94, 104, 81, 41, 40, 47, 20, 192,
-	66, 191, 119, 120, 80, 99, 190, 179, 99, 84,
-	178, 47, 102, 101, 47, 102, 101, 103, 177, 176,
-	103, 47, 175, 107, 117, 45, 164, 110, 65, 47,
-	105, 160, 163, 152, 129, 88, 83, 79, 77, 76,
-	116, 72, 118, 71, 70, 115, 60, 51, 125, 130,
-	131, 136, 189, 126, 188, 187, 137, 139, 140, 141,
-	142, 138, 181, 170, 149, 149, 143, 144, 145, 146,
-	150, 114, 47, 132, 159, 161, 162, 85, 73, 49,
-	158, 33, 157, 156, 75, 38, 68, 78, 86, 82,
-	39, 47, 47, 134, 32, 34, 171, 172, 173, 174,
-	35, 111, 128, 135, 47, 36, 112, 149, 180, 169,
-	154, 42, 63, 37, 87, 182, 113, 155, 69, 4,
-	183, 1, 97, 31, 93, 90, 133, 168, 153, 127,
-	43, 18, 17, 16, 15, 19, 14, 13, 12, 11,
-	10, 9, 8, 7, 6, 3, 5, 2,
+	109, 157, 104, 100, 98, 156, 105, 101, 67, 82,
+	131, 130, 176, 160, 174, 175, 22, 175, 23, 115,
+	133, 132, 24, 69, 25, 49, 26, 126, 195, 194,
+	47, 51, 54, 27, 28, 56, 57, 58, 59, 60,
+	61, 62, 63, 64, 193, 125, 66, 32, 107, 29,
+	124, 30, 118, 72, 31, 117, 113, 107, 103, 89,
+	50, 21, 74, 71, 48, 128, 129, 44, 43, 88,
+	201, 200, 108, 199, 188, 145, 93, 123, 50, 111,
+	110, 108, 187, 186, 112, 185, 184, 50, 111, 110,
+	50, 116, 70, 112, 173, 50, 119, 169, 114, 161,
+	138, 97, 92, 91, 73, 87, 85, 84, 80, 79,
+	78, 127, 77, 65, 55, 198, 197, 134, 139, 140,
+	196, 190, 135, 179, 94, 146, 148, 149, 150, 151,
+	147, 172, 141, 158, 158, 152, 153, 154, 155, 159,
+	52, 53, 81, 168, 170, 171, 167, 166, 165, 143,
+	36, 50, 50, 83, 50, 75, 41, 86, 90, 95,
+	120, 144, 42, 35, 37, 180, 181, 182, 183, 34,
+	38, 137, 121, 178, 50, 39, 158, 189, 163, 45,
+	68, 40, 96, 191, 122, 142, 164, 76, 192, 4,
+	1, 106, 102, 33, 99, 177, 162, 136, 46, 19,
+	18, 17, 16, 20, 15, 14, 13, 12, 11, 10,
+	9, 8, 7, 6, 3, 5, 2,
 }
 
 var yyPact = [...]int16{
 	7, -1000, -1000, 7, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
-	-1000, 119, 120, 153, 110, 5, 145, 33, 104, 9,
-	56, -1000, 9, 9, 9, 9, 9, 9, 9, 9,
-	-1000, 55, 9, 152, -1000, -1000, -31, -1000, 37, 9,
-	106, -1000, 169, 53, 52, 50, 91, 98, 48, 47,
-	-1000, 108, 46, 9, 2, -1000, 114, 45, 116, 112,
-	-1000, -1000, -1000, 156, 44, 23, -1000, -1000, 1, -1000,
-	98, -37, 9, -1000, -4, -12, 9, 127, -1000, -1000,
-	134, 166, -1000, 78, 23, 6, -1000, -1000, -33, -35,
-	-1000, -1000, -1000, 23, 152, 129, -1000, 43, 9, 103,
-	146, 61, 23, 23, 26, 26, 26, 26, -1000, 26,
-	26, 26, 26, 9, 9, -42, 42, 142, 165, -1000,
-	97, 96, 94, 41, 9, 84, 35, 166, -1000, 6,
-	6, 6, 6, -1000, -1000, -1000, -1000, -41, -1000, -1000,
-	-43, -1000, -1000, 138, 74, 9, 9, 9, 9, 31,
-	28, 27, 19, 16, -1000, -1000, 9, -1000, -1000, 73,
-	-1000, 164, -26, -27, -28, -1000, -1000, -1000, -1000, -1000,
-	-1000, -1000, -1000, -1000, 66, 65, 63, 15, 10, 8,
-	-1000, -1000, -1000,
+	-1000, -1000, 132, 126, 127, 159, 119, 14, 151, 9,
+	103, 39, 60, -1000, 39, 39, 39, 39, 39, 39,
+	39, 39, 39, -1000, 59, 39, 158, -1000, -1000, -34,
+	-1000, 38, 39, 50, 112, -1000, 178, 58, 56, 55,
+	54, 92, 104, 53, 52, -1000, 115, 51, 39, 4,
+	-1000, 121, 49, -1000, 48, 101, 120, -1000, -1000, -1000,
+	-1000, 162, 47, 27, -1000, -1000, 1, -1000, 104, -39,
+	39, -1000, -1000, -1, -4, 39, 124, -1000, -1000, 138,
+	174, -1000, 21, 27, 6, -1000, -1000, -36, -37, -1000,
+	-1000, -1000, 27, 158, 136, -1000, 46, 39, 100, 142,
+	22, 27, 27, 36, 36, 36, 36, -1000, 36, 36,
+	36, 36, 39, 39, -45, 45, 148, 172, -1000, 99,
+	98, 97, 44, 39, 123, 40, 174, -1000, 6, 6,
+	6, 6, -1000, -1000, -1000, -1000, -44, -1000, -1000, -46,
+	-1000, -1000, 140, 71, 39, 39, 39, 39, 32, 31,
+	29, 28, 20, -1000, -1000, 39, -1000, -1000, 69, -1000,
+	170, -12, -27, -28, -1000, -1000, -1000, -1000, -1000, -1000,
+	-1000, -1000, -1000, 68, 64, 63, 19, 17, 16, -1000,
+	-1000, -1000,
 }
 
 var yyPgo = [...]uint8{
-	0, 207, 206, 179, 205, 204, 203, 202, 201, 200,
-	199, 198, 197, 196, 195, 194, 193, 192, 191, 190,
-	8, 9, 189, 188, 187, 4, 185, 3, 184, 2,
-	7, 6, 182, 5, 1, 0, 181,
+	0, 216, 215, 189, 214, 213, 212, 211, 210, 209,
+	208, 207, 206, 205, 204, 203, 202, 201, 200, 199,
+	198, 8, 9, 197, 196, 195, 4, 194, 3, 192,
+	2, 7, 6, 191, 5, 1, 0, 190,
 }
 
 var yyR1 = [...]int8{
-	0, 36, 1, 4, 4, 3, 2, 2, 2, 2,
+	0, 37, 1, 4, 4, 3, 2, 2, 2, 2,
 	2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
-	2, 5, 5, 5, 5, 5, 6, 7, 8, 9,
-	10, 11, 12, 13, 13, 15, 14, 16, 17, 17,
-	17, 18, 18, 18, 19, 19, 19, 20, 21, 21,
-	22, 22, 22, 22, 23, 23, 24, 24, 25, 26,
-	26, 27, 27, 30, 30, 28, 28, 28, 28, 28,
-	29, 29, 29, 29, 29, 31, 31, 31, 32, 32,
-	32, 32, 33, 33, 34, 35,
+	2, 2, 6, 5, 5, 5, 5, 5, 7, 8,
+	9, 10, 11, 12, 13, 14, 14, 16, 15, 17,
+	18, 18, 18, 18, 18, 19, 19, 19, 20, 20,
+	20, 21, 22, 22, 23, 23, 23, 23, 24, 24,
+	25, 25, 26, 27, 27, 28, 28, 31, 31, 29,
+	29, 29, 29, 29, 30, 30, 30, 30, 30, 32,
+	32, 32, 33, 33, 33, 33, 34, 34, 35, 36,
 }
 
 var yyR2 = [...]int8{
 	0, 1, 1, 1, 2, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-	1, 9, 9, 9, 9, 9, 4, 4, 4, 8,
-	5, 4, 4, 2, 3, 7, 2, 4, 3, 6,
-	4, 11, 11, 11, 1, 1, 4, 6, 2, 0,
-	3, 4, 4, 0, 2, 0, 2, 0, 1, 1,
-	3, 1, 3, 1, 2, 1, 3, 3, 3, 3,
-	1, 3, 3, 3, 3, 1, 4, 4, 1, 1,
-	1, 3, 1, 3, 1, 1,
+	1, 1, 4, 9, 9, 9, 9, 9, 4, 4,
+	4, 8, 5, 4, 4, 2, 3, 7, 2, 4,
+	3, 6, 4, 3, 4, 11, 11, 11, 1, 1,
+	4, 6, 2, 0, 3, 4, 4, 0, 2, 0,
+	2, 0, 1, 1, 3, 1, 3, 1, 2, 1,
+	3, 3, 3, 3, 1, 3, 3, 3, 3, 1,
+	4, 4, 1, 1, 1, 3, 1, 3, 1, 1,
 }
 
 var yyChk = [...]int16{
-	-1000, -36, -1, -4, -3, -2, -5, -6, -7, -8,
-	-9, -10, -11, -12, -13, -15, -16, -17, -18, -14,
-	51, 9, 13, 15, 17, 24, 25, 39, 41, 44,
-	37, -3, 35, 21, 35, 40, 45, 20, 35, 40,
-	51, 50, 26, -19, -35, 52, 14, 48, -35, 35,
-	-35, 51, -35, -35, -35, -35, -35, -35, -35, -35,
-	51, -35, -20, 20, 54, 51, 23, -35, 40, 9,
-	51, 51, 51, 47, -21, 46, 51, 51, 39, 51,
-	-35, 52, 35, 51, -35, 21, 36, 18, 51, -25,
-	-26, -27, -30, -28, 29, -29, -31, -32, 19, 42,
-	-35, 50, 49, 54, 52, -21, 55, -35, 53, 53,
-	-35, 34, 32, 10, 53, 27, 22, 6, -30, 56,
-	57, 5, 4, 54, 54, -25, -20, -22, 33, 51,
-	-35, -35, 30, 40, 7, 17, 50, -27, -30, -29,
-	-29, -29, -29, -31, -31, -31, -31, -33, -34, -35,
-	-33, 55, 51, -23, 28, 12, 46, 46, 46, -35,
-	50, -35, -35, 8, 51, 55, 58, 55, -24, 31,
-	49, -35, -35, -35, -35, 51, 51, 51, 51, 51,
-	-34, 49, 11, 16, 53, 53, 53, 49, 49, 49,
-	51, 51, 51,
+	-1000, -37, -1, -4, -3, -2, -5, -6, -7, -8,
+	-9, -10, -11, -12, -13, -14, -16, -17, -18, -19,
+	-15, 54, 9, 11, 15, 17, 19, 26, 27, 42,
+	44, 47, 40, -3, 37, 37, 23, 37, 43, 48,
+	22, 37, 43, 54, 53, 28, -20, -36, 55, 16,
+	51, -36, 37, 38, -36, 54, -36, -36, -36, -36,
+	-36, -36, -36, -36, -36, 54, -36, -21, 22, 57,
+	54, 25, -36, 54, 12, 43, 9, 54, 54, 54,
+	54, 50, -22, 49, 54, 54, 42, 54, -36, 55,
+	37, 54, 54, -36, 23, 39, 20, 54, -26, -27,
+	-28, -31, -29, 31, -30, -32, -33, 21, 45, -36,
+	53, 52, 57, 55, -22, 58, -36, 56, 56, -36,
+	36, 34, 10, 56, 29, 24, 6, -31, 59, 60,
+	5, 4, 57, 57, -26, -21, -23, 35, 54, -36,
+	-36, 32, 43, 7, 19, 53, -28, -31, -30, -30,
+	-30, -30, -32, -32, -32, -32, -34, -35, -36, -34,
+	58, 54, -24, 30, 14, 49, 49, 49, -36, 53,
+	-36, -36, 8, 54, 58, 61, 58, -25, 33, 52,
+	-36, -36, -36, -36, 54, 54, 54, 54, 54, -35,
+	52, 13, 18, 56, 56, 56, 52, 52, 52, 54,
+	54, 54,
 }
 
 var yyDef = [...]int8{
 	0, -2, 1, 2, 3, 5, 6, 7, 8, 9,
 	10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
-	20, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-	0, 4, 0, 0, 0, 0, 0, 0, 0, 0,
-	33, 0, 0, 0, 44, 45, 0, 85, 0, 0,
-	0, 36, 0, 0, 0, 0, 0, 49, 0, 0,
-	34, 0, 0, 0, 0, 38, 0, 0, 0, 0,
-	26, 27, 28, 0, 0, 0, 31, 32, 0, 37,
-	49, 0, 0, 40, 0, 0, 0, 0, 30, 48,
-	58, 59, 61, 63, 0, 65, 70, 75, 0, 0,
-	78, 79, 80, 0, 0, 53, 46, 0, 0, 0,
-	0, 0, 0, 0, 0, 0, 0, 0, 64, 0,
-	0, 0, 0, 0, 0, 0, 0, 55, 0, 39,
-	0, 0, 0, 0, 0, 0, 0, 60, 62, 66,
-	67, 68, 69, 71, 72, 73, 74, 0, 82, 84,
-	0, 81, 35, 57, 0, 0, 0, 0, 0, 0,
-	0, 0, 0, 0, 29, 76, 0, 77, 47, 0,
-	54, 50, 0, 0, 0, 21, 22, 23, 24, 25,
-	83, 56, 51, 52, 0, 0, 0, 0, 0, 0,
-	41, 42, 43,
+	20, 21, 0, 0, 0, 0, 0, 0, 0, 0,
+	0, 0, 0, 4, 0, 0, 0, 0, 0, 0,
+	0, 0, 0, 35, 0, 0, 0, 48, 49, 0,
+	89, 0, 0, 0, 0, 38, 0, 0, 0, 0,
+	0, 0, 53, 0, 0, 36, 0, 0, 0, 0,
+	40, 0, 0, 43, 0, 0, 0, 22, 28, 29,
+	30, 0, 0, 0, 33, 34, 0, 39, 53, 0,
+	0, 42, 44, 0, 0, 0, 0, 32, 52, 62,
+	63, 65, 67, 0, 69, 74, 79, 0, 0, 82,
+	83, 84, 0, 0, 57, 50, 0, 0, 0, 0,
+	0, 0, 0, 0, 0, 0, 0, 68, 0, 0,
+	0, 0, 0, 0, 0, 0, 59, 0, 41, 0,
+	0, 0, 0, 0, 0, 0, 64, 66, 70, 71,
+	72, 73, 75, 76, 77, 78, 0, 86, 88, 0,
+	85, 37, 61, 0, 0, 0, 0, 0, 0, 0,
+	0, 0, 0, 31, 80, 0, 81, 51, 0, 58,
+	54, 0, 0, 0, 23, 24, 25, 26, 27, 87,
+	60, 55, 56, 0, 0, 0, 0, 0, 0, 45,
+	46, 47,
 }
 
 var yyTok1 = [...]int8{
@@ -270,9 +280,9 @@ var yyTok1 = [...]int8{
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
-	54, 55, 52, 3, 58, 3, 3, 3, 3, 3,
-	3, 3, 3, 3, 3, 3, 3, 3, 3, 51,
-	56, 53, 57,
+	57, 58, 55, 3, 61, 3, 3, 3, 3, 3,
+	3, 3, 3, 3, 3, 3, 3, 3, 3, 54,
+	59, 56, 60,
 }
 
 var yyTok2 = [...]int8{
@@ -280,7 +290,8 @@ var yyTok2 = [...]int8{
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
 	32, 33, 34, 35, 36, 37, 38, 39, 40, 41,
-	42, 43, 44, 45, 46, 47, 48, 49, 50,
+	42, 43, 44, 45, 46, 47, 48, 49, 50, 51,
+	52, 53,
 }
 
 var yyTok3 = [...]int8{
@@ -720,217 +731,217 @@ yydefault:
 	case 20:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
-			yyVAL.node = nil
+			yyVAL.node = yyDollar[1].node
 		}
 	case 21:
-		yyDollar = yyS[yypt-9 : yypt+1]
+		yyDollar = yyS[yypt-1 : yypt+1]
 		{
-			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Set, Value: yyDollar[8].str, StringLiteral: false}
+			yyVAL.node = nil
 		}
 	case 22:
-		yyDollar = yyS[yypt-9 : yypt+1]
+		yyDollar = yyS[yypt-4 : yypt+1]
 		{
-			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Set, Value: yyDollar[8].str, StringLiteral: true}
+			yyVAL.node = &ast.ArchiveProjectStmt{Project: yyDollar[3].str}
 		}
 	case 23:
 		yyDollar = yyS[yypt-9 : yypt+1]
 		{
-			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Add, Value: yyDollar[8].str}
+			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Set, Value: yyDollar[8].str, StringLiteral: false}
 		}
 	case 24:
 		yyDollar = yyS[yypt-9 : yypt+1]
 		{
-			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Drop, Value: yyDollar[8].str}
+			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Set, Value: yyDollar[8].str, StringLiteral: true}
 		}
 	case 25:
 		yyDollar = yyS[yypt-9 : yypt+1]
 		{
-			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Drop, Value: "*"}
+			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Add, Value: yyDollar[8].str}
 		}
 	case 26:
-		yyDollar = yyS[yypt-4 : yypt+1]
+		yyDollar = yyS[yypt-9 : yypt+1]
 		{
-			yyVAL.node = &ast.CreateFundStmt{Fund: yyDollar[3].str}
+			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Drop, Value: yyDollar[8].str}
 		}
 	case 27:
-		yyDollar = yyS[yypt-4 : yypt+1]
+		yyDollar = yyS[yypt-9 : yypt+1]
 		{
-			yyVAL.node = &ast.CreateProjectStmt{Project: yyDollar[3].str}
+			yyVAL.node = &ast.AlterProjectStmt{Project: yyDollar[3].str, Property: yyDollar[6].str, Action: ast.Drop, Value: "*"}
 		}
 	case 28:
 		yyDollar = yyS[yypt-4 : yypt+1]
 		{
-			yyVAL.node = &ast.CreateSetStmt{Set: yyDollar[3].str}
+			yyVAL.node = &ast.CreateFundStmt{Fund: yyDollar[3].str}
 		}
 	case 29:
+		yyDollar = yyS[yypt-4 : yypt+1]
+		{
+			yyVAL.node = &ast.CreateProjectStmt{Project: yyDollar[3].str}
+		}
+	case 30:
+		yyDollar = yyS[yypt-4 : yypt+1]
+		{
+			yyVAL.node = &ast.CreateSetStmt{Set: yyDollar[3].str}
+		}
+	case 31:
 		yyDollar = yyS[yypt-8 : yypt+1]
 		{
 			yyVAL.node = &ast.CreateUserStmt{User: yyDollar[3].str, EncryptedPassword: yyDollar[7].str}
 		}
-	case 30:
+	case 32:
 		yyDollar = yyS[yypt-5 : yypt+1]
 		{
 			yyVAL.node = &ast.DeleteStmt{From: yyDollar[3].str, Where: yyDollar[4].node}
 		}
-	case 31:
+	case 33:
 		yyDollar = yyS[yypt-4 : yypt+1]
 		{
 			yyVAL.node = &ast.DropProjectStmt{Project: yyDollar[3].str}
 		}
-	case 32:
+	case 34:
 		yyDollar = yyS[yypt-4 : yypt+1]
 		{
 			yyVAL.node = &ast.DropSetStmt{Set: yyDollar[3].str}
 		}
-	case 33:
+	case 35:
 		yyDollar = yyS[yypt-2 : yypt+1]
 		{
 			yyVAL.node = &ast.InfoStmt{Topic: ""}
 		}
-	case 34:
+	case 36:
 		yyDollar = yyS[yypt-3 : yypt+1]
 		{
 			yyVAL.node = &ast.InfoStmt{Topic: yyDollar[2].str}
 		}
-	case 35:
+	case 37:
 		yyDollar = yyS[yypt-7 : yypt+1]
 		{
 			yyVAL.node = &ast.InsertStmt{Into: yyDollar[3].str, Query: yyDollar[6].node}
 		}
-	case 36:
+	case 38:
 		yyDollar = yyS[yypt-2 : yypt+1]
 		{
 			yyVAL.node = &ast.PingStmt{}
 		}
-	case 37:
+	case 39:
 		yyDollar = yyS[yypt-4 : yypt+1]
 		{
 			yyVAL.node = &ast.SelectStmt{AttrList: yyDollar[2].node, Query: yyDollar[3].node}
 		}
-	case 38:
+	case 40:
 		yyDollar = yyS[yypt-3 : yypt+1]
 		{
 			yyVAL.node = &ast.ShowStmt{Type: yyDollar[2].str}
 		}
-	case 39:
+	case 41:
 		yyDollar = yyS[yypt-6 : yypt+1]
 		{
 			yyVAL.node = &ast.ShowStmt{Type: yyDollar[2].str, In: yyDollar[5].str}
 		}
-	case 40:
+	case 42:
 		yyDollar = yyS[yypt-4 : yypt+1]
 		{
 			yyVAL.node = &ast.ShowStmt{Type: "project", Name: yyDollar[3].str}
 		}
-	case 41:
+	case 43:
+		yyDollar = yyS[yypt-3 : yypt+1]
+		{
+			yyVAL.node = &ast.ShowStmt{Type: "projects"}
+		}
+	case 44:
+		yyDollar = yyS[yypt-4 : yypt+1]
+		{
+			yyVAL.node = &ast.ShowStmt{Type: "projects", Archived: true}
+		}
+	case 45:
 		yyDollar = yyS[yypt-11 : yypt+1]
 		{
 			yyVAL.node = &ast.UpdateStmt{Set: yyDollar[2].str, Attr: yyDollar[4].str, Value: yyDollar[6].str, StringLiteral: false, IDAttr: yyDollar[8].str, IDValue: &ast.Number{Value: yyDollar[10].str}}
 		}
-	case 42:
+	case 46:
 		yyDollar = yyS[yypt-11 : yypt+1]
 		{
 			yyVAL.node = &ast.UpdateStmt{Set: yyDollar[2].str, Attr: "fund", Value: yyDollar[6].str, StringLiteral: false, IDAttr: yyDollar[8].str, IDValue: &ast.Number{Value: yyDollar[10].str}}
 		}
-	case 43:
+	case 47:
 		yyDollar = yyS[yypt-11 : yypt+1]
 		{
 			yyVAL.node = &ast.UpdateStmt{Set: yyDollar[2].str, Attr: "fund", Value: "", StringLiteral: false, IDAttr: yyDollar[8].str, IDValue: &ast.Number{Value: yyDollar[10].str}}
 		}
-	case 44:
+	case 48:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
 			yyVAL.node = &ast.SelectAttrList{Attr: yyDollar[1].str}
 		}
-	case 45:
+	case 49:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
 			yyVAL.node = &ast.SelectAttrList{Attr: "*"}
 		}
-	case 46:
+	case 50:
 		yyDollar = yyS[yypt-4 : yypt+1]
 		{
 			yyVAL.node = &ast.SelectAttrList{Attr: "count(*)"}
 		}
-	case 47:
+	case 51:
 		yyDollar = yyS[yypt-6 : yypt+1]
 		{
 			yyVAL.node = &ast.QueryClause{From: yyDollar[2].str, Where: yyDollar[3].node, Order: yyDollar[4].node, Limit: yyDollar[5].node, Offset: yyDollar[6].node}
 		}
-	case 48:
+	case 52:
 		yyDollar = yyS[yypt-2 : yypt+1]
 		{
 			yyVAL.node = &ast.WhereClause{Valid: true, Condition: yyDollar[2].node}
 		}
-	case 49:
+	case 53:
 		yyDollar = yyS[yypt-0 : yypt+1]
 		{
 			yyVAL.node = &ast.WhereClause{}
 		}
-	case 50:
+	case 54:
 		yyDollar = yyS[yypt-3 : yypt+1]
 		{
 			yyVAL.node = &ast.OrderClause{Valid: true, Attr: yyDollar[3].str, Desc: false}
 		}
-	case 51:
+	case 55:
 		yyDollar = yyS[yypt-4 : yypt+1]
 		{
 			yyVAL.node = &ast.OrderClause{Valid: true, Attr: yyDollar[3].str, Desc: false}
 		}
-	case 52:
+	case 56:
 		yyDollar = yyS[yypt-4 : yypt+1]
 		{
 			yyVAL.node = &ast.OrderClause{Valid: true, Attr: yyDollar[3].str, Desc: true}
 		}
-	case 53:
+	case 57:
 		yyDollar = yyS[yypt-0 : yypt+1]
 		{
 			yyVAL.node = &ast.OrderClause{}
 		}
-	case 54:
+	case 58:
 		yyDollar = yyS[yypt-2 : yypt+1]
 		{
 			yyVAL.node = &ast.LimitClause{Valid: true, Count: yyDollar[2].str}
 		}
-	case 55:
+	case 59:
 		yyDollar = yyS[yypt-0 : yypt+1]
 		{
 			yyVAL.node = &ast.LimitClause{}
 		}
-	case 56:
+	case 60:
 		yyDollar = yyS[yypt-2 : yypt+1]
 		{
 			yyVAL.node = &ast.OffsetClause{Valid: true, Start: yyDollar[2].str}
 		}
-	case 57:
+	case 61:
 		yyDollar = yyS[yypt-0 : yypt+1]
 		{
 			yyVAL.node = &ast.OffsetClause{}
 		}
-	case 58:
-		yyDollar = yyS[yypt-1 : yypt+1]
-		{
-			yyVAL.node = yyDollar[1].node
-		}
-	case 59:
-		yyDollar = yyS[yypt-1 : yypt+1]
-		{
-			yyVAL.node = yyDollar[1].node
-		}
-	case 60:
-		yyDollar = yyS[yypt-3 : yypt+1]
-		{
-			yyVAL.node = &ast.OrExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
-		}
-	case 61:
-		yyDollar = yyS[yypt-1 : yypt+1]
-		{
-			yyVAL.node = yyDollar[1].node
-		}
 	case 62:
-		yyDollar = yyS[yypt-3 : yypt+1]
+		yyDollar = yyS[yypt-1 : yypt+1]
 		{
-			yyVAL.node = &ast.AndExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
+			yyVAL.node = yyDollar[1].node
 		}
 	case 63:
 		yyDollar = yyS[yypt-1 : yypt+1]
@@ -938,9 +949,9 @@ yydefault:
 			yyVAL.node = yyDollar[1].node
 		}
 	case 64:
-		yyDollar = yyS[yypt-2 : yypt+1]
+		yyDollar = yyS[yypt-3 : yypt+1]
 		{
-			yyVAL.node = &ast.NotExpr{Expr: yyDollar[2].node}
+			yyVAL.node = &ast.OrExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 65:
 		yyDollar = yyS[yypt-1 : yypt+1]
@@ -950,99 +961,119 @@ yydefault:
 	case 66:
 		yyDollar = yyS[yypt-3 : yypt+1]
 		{
-			yyVAL.node = &ast.EqualExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
+			yyVAL.node = &ast.AndExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 67:
-		yyDollar = yyS[yypt-3 : yypt+1]
-		{
-			yyVAL.node = &ast.LikeExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
-		}
-	case 68:
-		yyDollar = yyS[yypt-3 : yypt+1]
-		{
-			yyVAL.node = &ast.ILikeExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
-		}
-	case 69:
-		yyDollar = yyS[yypt-3 : yypt+1]
-		{
-			yyVAL.node = &ast.NotEqualExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
-		}
-	case 70:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
 			yyVAL.node = yyDollar[1].node
+		}
+	case 68:
+		yyDollar = yyS[yypt-2 : yypt+1]
+		{
+			yyVAL.node = &ast.NotExpr{Expr: yyDollar[2].node}
+		}
+	case 69:
+		yyDollar = yyS[yypt-1 : yypt+1]
+		{
+			yyVAL.node = yyDollar[1].node
+		}
+	case 70:
+		yyDollar = yyS[yypt-3 : yypt+1]
+		{
+			yyVAL.node = &ast.EqualExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 71:
 		yyDollar = yyS[yypt-3 : yypt+1]
 		{
-			yyVAL.node = &ast.LessThanExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
+			yyVAL.node = &ast.LikeExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 72:
 		yyDollar = yyS[yypt-3 : yypt+1]
 		{
-			yyVAL.node = &ast.GreaterThanExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
+			yyVAL.node = &ast.ILikeExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 73:
 		yyDollar = yyS[yypt-3 : yypt+1]
 		{
-			yyVAL.node = &ast.LessThanOrEqualExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
+			yyVAL.node = &ast.NotEqualExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 74:
-		yyDollar = yyS[yypt-3 : yypt+1]
-		{
-			yyVAL.node = &ast.GreaterThanOrEqualExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
-		}
-	case 75:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
 			yyVAL.node = yyDollar[1].node
 		}
-	case 76:
-		yyDollar = yyS[yypt-4 : yypt+1]
+	case 75:
+		yyDollar = yyS[yypt-3 : yypt+1]
 		{
-			yyVAL.node = &ast.FilterExpr{ExprList: yyDollar[3].nodeList}
+			yyVAL.node = &ast.LessThanExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
+		}
+	case 76:
+		yyDollar = yyS[yypt-3 : yypt+1]
+		{
+			yyVAL.node = &ast.GreaterThanExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 77:
-		yyDollar = yyS[yypt-4 : yypt+1]
+		yyDollar = yyS[yypt-3 : yypt+1]
 		{
-			yyVAL.node = &ast.TagExpr{ExprList: yyDollar[3].nodeList}
+			yyVAL.node = &ast.LessThanOrEqualExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 78:
-		yyDollar = yyS[yypt-1 : yypt+1]
+		yyDollar = yyS[yypt-3 : yypt+1]
 		{
-			yyVAL.node = &ast.Name{Value: yyDollar[1].str}
+			yyVAL.node = &ast.GreaterThanOrEqualExpr{Expr1: yyDollar[1].node, Expr2: yyDollar[3].node}
 		}
 	case 79:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
-			yyVAL.node = &ast.SLiteral{Value: yyDollar[1].str}
+			yyVAL.node = yyDollar[1].node
 		}
 	case 80:
-		yyDollar = yyS[yypt-1 : yypt+1]
+		yyDollar = yyS[yypt-4 : yypt+1]
 		{
-			yyVAL.node = &ast.Number{Value: yyDollar[1].str}
+			yyVAL.node = &ast.FilterExpr{ExprList: yyDollar[3].nodeList}
 		}
 	case 81:
-		yyDollar = yyS[yypt-3 : yypt+1]
+		yyDollar = yyS[yypt-4 : yypt+1]
 		{
-			yyVAL.node = &ast.ParenExpr{Expr: yyDollar[2].node}
+			yyVAL.node = &ast.TagExpr{ExprList: yyDollar[3].nodeList}
 		}
 	case 82:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
-			yyVAL.nodeList = yyDollar[1].nodeList
+			yyVAL.node = &ast.Name{Value: yyDollar[1].str}
 		}
 	case 83:
-		yyDollar = yyS[yypt-3 : yypt+1]
+		yyDollar = yyS[yypt-1 : yypt+1]
 		{
-			yyVAL.nodeList = append(yyDollar[1].nodeList, yyDollar[3].nodeList...)
+			yyVAL.node = &ast.SLiteral{Value: yyDollar[1].str}
 		}
 	case 84:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
-			yyVAL.nodeList = []ast.Node{&ast.Name{Value: yyDollar[1].str}}
+			yyVAL.node = &ast.Number{Value: yyDollar[1].str}
 		}
 	case 85:
+		yyDollar = yyS[yypt-3 : yypt+1]
+		{
+			yyVAL.node = &ast.ParenExpr{Expr: yyDollar[2].node}
+		}
+	case 86:
+		yyDollar = yyS[yypt-1 : yypt+1]
+		{
+			yyVAL.nodeList = yyDollar[1].nodeList
+		}
+	case 87:
+		yyDollar = yyS[yypt-3 : yypt+1]
+		{
+			yyVAL.nodeList = append(yyDollar[1].nodeList, yyDollar[3].nodeList...)
+		}
+	case 88:
+		yyDollar = yyS[yypt-1 : yypt+1]
+		{
+			yyVAL.nodeList = []ast.Node{&ast.Name{Value: yyDollar[1].str}}
+		}
+	case 89:
 		yyDollar = yyS[yypt-1 : yypt+1]
 		{
 			yyVAL.str = strings.ToLower(yyDollar[1].str)