+++ cmd/ccd/server/createuser.go
@@ -11,7 +11,7 @@ import (
 func createUserStmt(s *svr, rqid int64, cmd *ast.CreateUserStmt) *ccms.Result {
 	userExists, err := cat.UserExists(s.d, cmd.User)
 	if err != nil {
-		return cmderrint("checking if user exists", err)
+		return cmderr("checking if user exists: " + err.Error())
 	}
 	if userExists {
 		return cmderr("user \"" + cmd.User + "\" already exists")
@@ -22,7 +22,7 @@ func createUserStmt(s *svr, rqid int64, cmd *ast.CreateUserStmt) *ccms.Result {
 	}
 
 	if err := cat.CreateUser(s.conf.Security.SecretKey, s.d, cmd.User, cmd.EncryptedPassword, false, true); err != nil {
-		return cmderrint("writing user", err)
+		return cmderr("writing user: " + err.Error())
 	}
 
 	return ccms.NewResult("create user")